Change Console output to cast entries as double

+++ ConsoleScanner/Program.cs
@@ -48,6 +48,6 @@ Console.WriteLine("Id available:");
 sql = $@"SELECT id FROM {tableName}";
 
 var entries = dataAccess.ReadData<dynamic>(sql);
-entries.ForEach(d => Console.WriteLine(d));
+entries.ForEach(d => Console.WriteLine((double)d));
 
 return;
